add response helper and fix api contrllers

+++ app/Http/Controllers/EventController.php
@@ -5,7 +5,11 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Event\StoreEventRequest;
 use App\Http\Requests\Event\UpdateEventRequest;
 use App\Models\Event;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response as ResponseAlias;
+use function App\Helpers\serverErrorResponse;
+use function App\Helpers\swcResponse;
 
 class EventController extends Controller
 {
@@ -14,29 +18,47 @@ class EventController extends Controller
      */
     public function index()
     {
-        return Event::all();
+        return swcResponse(Event::all()->toArray());
     }
 
-    public function store(StoreEventRequest $request): Event
+    public function store(StoreEventRequest $request): JsonResponse
     {
         $data = $request->validated();
-        return Event::create($data);
-    }
-    public function show(Event $event)
-    {
-        return $event;
+        try {
+            $event = Event::create($data);
+            return swcResponse($event->toArray());
+        }catch (\Throwable $error) {
+            return serverErrorResponse($error->getMessage());
+        }
+
     }
 
-    public function update(UpdateEventRequest $request, Event $event): Event
+    public function show(Event $event): JsonResponse
+    {
+        return swcResponse($event->toArray());
+    }
+
+    public function update(UpdateEventRequest $request, Event $event): JsonResponse
     {
         $data = $request->validated();
-        $event = $event->update($data);
-        return new Event($event);
+        try {
+            $event->update($data);
+        } catch (\Throwable $error) {
+            return serverErrorResponse($error->getMessage());
+        }
+        return swcResponse($event->toArray());
     }
 
-    public function destroy(Event $event)
+    public function destroy(Event $event): JsonResponse
     {
-        $event->deleteOrFail();
-        return response(null, ResponseAlias::HTTP_NO_CONTENT);
+        if ($event->getAttribute('creator_id') !== Auth::user()->id) {
+            return swcResponse([], ResponseAlias::HTTP_FORBIDDEN, 'access denied');
+        }
+        try {
+            $event->deleteOrFail();
+        }catch (\Throwable $error) {
+            return serverErrorResponse($error->getMessage());
+        }
+        return swcResponse([], ResponseAlias::HTTP_NO_CONTENT);
     }
 }
